requirements.txt
pytz
pandas
#fastapi
python-dotenv
pydantic
pikepdf
#langchain
#openpyxl
#tiktoken
#openai
#faiss-cpu
#pip-tools
google-cloud-storage
google-oauth2-tool
google-cloud-bigquery
google-auth
pandas-gbq
#openpyxl